refactor(forges): decompose jira.rs into jira/ submodule
Split the 1982-line forges/jira.rs into domain-focused submodules:
- jira/oauth.rs: OAuth PKCE flow and credential management
- jira/types.rs: API response types
- jira/adf.rs: Atlassian Document Format conversion
- jira/client.rs: JiraClient struct and HTTP methods
- jira/mod.rs: Re-exports, auth config, link flow, Forge impl

Follows the same screaming architecture pattern as github/ and linear/.
